Added Recursive Call for ftp_put
Changed ftp put method to recursively walk the directory structure of
the specified folder (base) and upload and create the directory
structure on the receiving end while it uploads.

This is useful for when you dont want to stuff all the images into a
single directory, and would prefer to maintain the YYYY/MM structure
that WordPress gives you.